--- src/main/java/com/dev/smartcash/offer/OfferRepository.java
@@ -1,5 +1,0 @@
-package com.dev.smartcash.offer;
-
-public interface OfferRepository {
-    Offer save(Offer offer);
-}
