--- src/main/java/com/dev/smartcash/offer/OfferService.java
@@ -1,5 +1,0 @@
-package com.dev.smartcash.offer;
-
-public interface OfferService {
-    OfferResponse newOffer(OfferRequest offerRequest);
-}
