+++ src/main/java/com/dev/smartcash/offer/application/repository/OfferRepository.java
@@ -1,0 +1,7 @@
+package com.dev.smartcash.offer.application.repository;
+
+import com.dev.smartcash.offer.domain.Offer;
+
+public interface OfferRepository {
+    Offer save(Offer offer);
+}
