+++ src/main/java/com/dev/smartcash/offer/application/service/OfferService.java
@@ -1,0 +1,8 @@
+package com.dev.smartcash.offer.application.service;
+
+import com.dev.smartcash.offer.application.api.OfferRequest;
+import com.dev.smartcash.offer.application.api.OfferResponse;
+
+public interface OfferService {
+    OfferResponse newOffer(OfferRequest offerRequest);
+}
