+++ src/main/java/com/dev/smartcash/offer/infra/OfferInfraRepository.java
@@ -1,5 +1,7 @@
-package com.dev.smartcash.offer;
+package com.dev.smartcash.offer.infra;
 
+import com.dev.smartcash.offer.application.repository.OfferRepository;
+import com.dev.smartcash.offer.domain.Offer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Repository;
@@ -7,7 +9,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 @Log4j2
 @RequiredArgsConstructor
-public class OfferInfraRepository implements OfferRepository{
+public class OfferInfraRepository implements OfferRepository {
 
     private final OfferSpringDataJPARepository offerSpringDataJPARepository;
